backend/app/requirements.txt
fastapi
uvicorn
transformers  # For Hugging Face models
tensorflow    # If using TensorFlow models
pillow        # For image processing
python-dotenv # For loading environment variables

